Created hard coded data for the restaurant details
I integrated the css from the index and the restaurant details for the styling to use the common css and then make adjustments for certain tags. once the database is setup we will have to go in and change the hard coded data to be able to change based on the restaurant.