Connected Top 4 restaurants and Restaurant list  to backend
You are able to search "the top 4" restaurants on the home page by putting in a zip code.
You are able to search for restaurants by giving a zip code on restaurant list, however if you click on a restaurant for restaurant details page no data is generated.

Also no images are generated if there is not a image for that location in our database